feat(ProfileScreen): add Work History and Feedback section
-Work History and Feedback section completed
-ProfileScreen code structure updated
-code cleanup
