+++ app/src/main/java/com/dyshuk/android/upworkredesigncompose/ui/screens/profile_screen/ProfileViewModel.kt
@@ -3,6 +3,9 @@ package com.dyshuk.android.upworkredesigncompose.ui.screens.profile_screen
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.dyshuk.android.upworkredesigncompose.data.repository.FakeFeedbackRepository
+import com.dyshuk.android.upworkredesigncompose.ui.screens.profile_screen.ui_state.FeedbackListState
+import com.dyshuk.android.upworkredesigncompose.ui.screens.profile_screen.ui_state.ProfileScreenState
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -20,6 +23,7 @@ class ProfileViewModel() : ViewModel() {
 
     private fun fetchFeedbackList() {
         viewModelScope.launch {
+            delay(1200)
             _uiState.update { it.copy(feedbackListState = FeedbackListState.Loading) }
             try {
                 val feedbackList = FakeFeedbackRepository.getFeedbackList()
