+++ app/src/main/java/com/dyshuk/android/upworkredesigncompose/ui/screens/profile_screen/ui_state/JobScreenState.kt
@@ -1,0 +1,6 @@
+package com.dyshuk.android.upworkredesigncompose.ui.screens.profile_screen.ui_state
+
+data class ProfileScreenState(
+    val feedbackListState: FeedbackListState = FeedbackListState.Loading
+)
+
