+++ gradle/libs.versions.toml
@@ -1,5 +1,5 @@
 [versions]
-agp = "8.10.1"
+agp = "8.11.0"
 coreSplashscreen = "1.0.1"
 kotlin = "2.1.0"
 ktlint = "12.1.1"
